split batch and standard tests
