+++ tests/core.py
@@ -1,5 +1,3 @@
-from pathlib import Path
-
 import pytest
 import snowflake.sqlalchemy.custom_types as sct
 import sqlalchemy
@@ -21,7 +19,6 @@ from singer_sdk.testing.target_tests import (
     TargetSchemaUpdates,
     TargetSpecialCharsInAttributes,
 )
-from singer_sdk.testing.templates import TargetFileTestTemplate
 
 
 class SnowflakeTargetArrayData(TargetArrayData):
@@ -236,175 +233,27 @@ class SnowflakeTargetSchemaUpdates(TargetSchemaUpdates):
             isinstance(column.type, expected_types[column.name])
 
 
-# BATCH Tests
-
-
-class SnowflakeTargetCustomTestTemplate(TargetFileTestTemplate):
-    @property
-    def singer_filepath(self) -> Path:
-        """Get path to singer JSONL formatted messages file.
-
-        Files will be sourced from `./target_test_streams/<test name>.singer`.
-
-        Returns:
-            The expected Path to this tests singer file.
-        """
-        current_dir = Path(__file__).resolve().parent
-        return current_dir / "target_test_streams" / f"{self.name}.singer"
-
-
-class SnowflakeTargetBatchArrayData(
-    SnowflakeTargetCustomTestTemplate, SnowflakeTargetArrayData
-):
-    """Test that the target can handle batch messages."""
-
-    name = "batch_array_data"
-
-
-class SnowflakeTargetBatchCamelcase(
-    SnowflakeTargetCustomTestTemplate, SnowflakeTargetCamelcaseTest
-):
-    """Test that the target can handle batch messages."""
-
-    name = "batch_camelcase"
-
-
-class SnowflakeTargetBatchDuplicateRecords(
-    SnowflakeTargetCustomTestTemplate, SnowflakeTargetDuplicateRecords
-):
-    """Test that the target can handle batch messages."""
-
-    name = "batch_duplicate_records"
-
-
-class SnowflakeTargetBatchEncodedStringData(
-    SnowflakeTargetCustomTestTemplate, SnowflakeTargetEncodedStringData
-):
-    """Test that the target can handle batch messages."""
-
-    name = "batch_encoded_string_data"
-
-
-# class SnowflakeTargetBatchMultipleStateMessages(
-#     SnowflakeTargetCustomTestTemplate, TargetMultipleStateMessages
-# ):
-#     """Test that the target can handle batch messages."""
-
-#     name = "batch_multiple_state_messages"
-
-
-# class SnowflakeTargetBatchNoPrimaryKeysAppend(
-#     SnowflakeTargetCustomTestTemplate, TargetNoPrimaryKeysAppend
-# ):
-#     """Test that the target can handle batch messages."""
-
-#     name = "batch_no_primary_keys_append"
-
-
-class SnowflakeTargetBatchNoPrimaryKeys(
-    SnowflakeTargetCustomTestTemplate, TargetNoPrimaryKeys
-):
-    """Test that the target can handle batch messages."""
-
-    name = "batch_no_primary_keys"
-
-
-class SnowflakeTargetBatchOptionalAttributes(
-    SnowflakeTargetCustomTestTemplate, TargetOptionalAttributes
-):
-    """Test that the target can handle batch messages."""
-
-    name = "batch_optional_attributes"
-
-
-class SnowflakeTargetBatchRecordBeforeSchemaTest(
-    SnowflakeTargetCustomTestTemplate, SnowflakeTargetRecordBeforeSchemaTest
-):
-    """Test that the target can handle batch messages."""
-
-    name = "batch_record_before_schema"
-
-
-class SnowflakeTargetBatchRecordMissingKeyProperty(
-    SnowflakeTargetCustomTestTemplate, SnowflakeTargetRecordMissingKeyProperty
-):
-    """Test that the target can handle batch messages."""
-
-    name = "batch_record_missing_key_property"
-
-
-class SnowflakeTargetBatchRecordMissingRequiredProperty(
-    SnowflakeTargetCustomTestTemplate, TargetRecordMissingRequiredProperty
-):
-    """Test that the target can handle batch messages."""
-
-    name = "batch_record_missing_required_property"
-
-
-class SnowflakeTargetBatchSchemaNoProperties(
-    SnowflakeTargetCustomTestTemplate, SnowflakeTargetSchemaNoProperties
-):
-    """Test that the target can handle batch messages."""
-
-    name = "batch_schema_no_properties"
-
-
-class SnowflakeTargetBatchSchemaUpdates(
-    SnowflakeTargetCustomTestTemplate, SnowflakeTargetSchemaUpdates
-):
-    """Test that the target can handle batch messages."""
-
-    name = "batch_schema_updates"
-
-
-class SnowflakeTargetBatchSpecialCharsInAttributes(
-    SnowflakeTargetCustomTestTemplate, TargetSpecialCharsInAttributes
-):
-    """Test that the target can handle batch messages."""
-
-    name = "batch_special_chars_in_attributes"
-
-
 target_tests = TestSuite(
     kind="target",
     tests=[
         # Core
-        # SnowflakeTargetArrayData,
-        # SnowflakeTargetCamelcaseComplexSchema,
-        # SnowflakeTargetCamelcaseTest,
-        # TargetCliPrintsTest,
-        # # TODO: bug https://github.com/MeltanoLabs/target-snowflake/issues/41
-        # # SnowflakeTargetDuplicateRecords,
-        # SnowflakeTargetEncodedStringData,
-        # SnowflakeTargetInvalidSchemaTest,
-        # # Not available in the SDK yet
-        # # TargetMultipleStateMessages,
-        # TargetNoPrimaryKeys,  # Implicitly asserts no pk is handled
-        # TargetOptionalAttributes,  # Implicitly asserts that nullable fields are handled
-        # SnowflakeTargetRecordBeforeSchemaTest,
-        # SnowflakeTargetRecordMissingKeyProperty,
-        # TargetRecordMissingRequiredProperty,
-        # SnowflakeTargetSchemaNoProperties,
-        # SnowflakeTargetSchemaUpdates,
-        # TargetSpecialCharsInAttributes,  # Implicitly asserts that special chars are handled
-        # BATCH
-        # TODO: failing assertion
-        # SnowflakeTargetBatchArrayData,
-        SnowflakeTargetBatchCamelcase,
-        # TODO: as above
-        # SnowflakeTargetBatchDuplicateRecords,
-        SnowflakeTargetBatchEncodedStringData,
-        # TODO: as above
-        # SnowflakeTargetBatchMultipleStateMessages,
-        # TODO: in sdk but not enabled
-        # SnowflakeTargetBatchNoPrimaryKeysAppend,
-        SnowflakeTargetBatchNoPrimaryKeys,
-        # TODO: failing on SQL Merge (NULLs in non-Nullable columns)
-        # SnowflakeTargetBatchOptionalAttributes,
-        SnowflakeTargetBatchRecordBeforeSchemaTest,
-        SnowflakeTargetBatchRecordMissingKeyProperty,
-        SnowflakeTargetBatchSchemaNoProperties,
-        SnowflakeTargetBatchSchemaUpdates,
-        SnowflakeTargetBatchSpecialCharsInAttributes,
+        SnowflakeTargetArrayData,
+        SnowflakeTargetCamelcaseComplexSchema,
+        SnowflakeTargetCamelcaseTest,
+        TargetCliPrintsTest,
+        # TODO: bug https://github.com/MeltanoLabs/target-snowflake/issues/41
+        # SnowflakeTargetDuplicateRecords,
+        SnowflakeTargetEncodedStringData,
+        SnowflakeTargetInvalidSchemaTest,
+        # Not available in the SDK yet
+        # TargetMultipleStateMessages,
+        TargetNoPrimaryKeys,  # Implicitly asserts no pk is handled
+        TargetOptionalAttributes,  # Implicitly asserts that nullable fields are handled
+        SnowflakeTargetRecordBeforeSchemaTest,
+        SnowflakeTargetRecordMissingKeyProperty,
+        TargetRecordMissingRequiredProperty,
+        SnowflakeTargetSchemaNoProperties,
+        SnowflakeTargetSchemaUpdates,
+        TargetSpecialCharsInAttributes,  # Implicitly asserts that special chars are handled
     ],
 )
