--- tests/target_test_streams/user_location_upsert_data__locations.jsonl
@@ -1,4 +1,0 @@
-{"type": "RECORD", "stream": "test_locations", "record": {"id": 1, "name": "Philly"}}
-{"type": "RECORD", "stream": "test_locations", "record": {"id": 3, "name": "San Francisco"}}
-{"type": "RECORD", "stream": "test_locations", "record": {"id": 6, "name": "Colorado"}}
-{"type": "RECORD", "stream": "test_locations", "record": {"id": 8, "name": "Boston"}}
